chore: centralise charm fixtures
